Translate the base case
Translate scores created and downloaded from Sequencer. You cannot then upload the score, download again, regardless of changes, as the tempo field changes from an int to a string and chords can include TEMPO=N as a string mixed with int note values.

Still have to handle this uploaded/downloaded case, but this is the first working translation step.